apply suggested changes

+++ src/main/java/edu/hw1/Task1.java
@@ -2,10 +2,10 @@ package edu.hw1;
 
 public class Task1 {
     @SuppressWarnings("MagicNumber")
-    public int minutesToSeconds(String time) {
+    public long minutesToSeconds(String time) {
         String[] params = time.split(":");
         int seconds = Integer.parseInt(params[1]);
-        int minutes = Integer.parseInt(params[0]);
+        long minutes = Long.parseLong(params[0]);
 
         if (seconds >= 60 || seconds < 0) {
             throw new IllegalArgumentException("Seconds expected to be in range [0;60)");
